api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.136.1",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.1",
    "uvicorn>=0.46.0",
]
